Better National Themes
Shortened the themes, around 30 seconds each, for less storage consumption, better quality, and just because it's better, having played a lot with the old themes it got a bit excesive when they were so long. And, removed the national Anthem, because altough we like it very much, it represents more the nation, rather than a culture, a civilization.

Also added themes for Cuba and Panama, and, gladly changed the theme songs for Bolivia and El Salvador for better representation.